main.m
This code solves the Lotka-Volterra system of ordinary differential equations and plots the corresponding graphs. To do this, the initial conditions and parameters that control the dynamics of the population (r0 and f0) are set, and the system of differential equations is solved using the `ode45` function.

Next, the mean error and standard deviation are calculated for the result based on 1000 random sets of initial conditions and parameters `r0` and `f0`.